requirements.txt
# ── Core data stack ─────────────────────────────────────────────────
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
pyarrow==15.0.0

# ── Ethereum / blockchain ───────────────────────────────────────────
web3==6.15.1
requests==2.31.0
aiohttp==3.9.3
tenacity==8.2.3

# ── Machine learning ────────────────────────────────────────────────
scikit-learn==1.4.1.post1
lightgbm==4.3.0
shap==0.45.0
imbalanced-learn==0.12.0

# ── Visualization ───────────────────────────────────────────────────
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0

# ── Statistical analysis (credit risk metrics) ──────────────────────
scipy==1.12.0
statsmodels==0.14.1

# ── Notebooks ───────────────────────────────────────────────────────
jupyter==1.0.0
ipykernel==6.29.3
notebook==7.1.1

# ── API (Phase 3) ───────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3

# ── Database ────────────────────────────────────────────────────────
sqlalchemy==2.0.27
psycopg2-binary==2.9.9

# ── Testing ─────────────────────────────────────────────────────────
pytest==8.0.2
pytest-asyncio==0.23.5

# ── Code quality ────────────────────────────────────────────────────
black==24.2.0
ruff==0.3.0
